requirements.txt
Flask==2.1.1
python-docx==0.8.11
PyPDF2==1.26.0
pandas==1.3.3
Werkzeug==2.0.2